Moved value evaluator into analysis package
