+++ cpg-library/src/test/resources/value_evaluation/example.cpp
